Refactored ultragraph's trait system and algo implementations for better maintainability.
