+++ ultragraph/src/traits/graph_algo_pathfinder.rs
@@ -1,0 +1,39 @@
+/*
+ * SPDX-License-Identifier: MIT
+ * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
+ */
+use crate::{GraphError, GraphView};
+
+pub trait PathfindingGraphAlgorithms<N, W>: GraphView<N, W> {
+    /// Checks if a path of any length exists from a start to a stop index.
+    fn is_reachable(&self, start_index: usize, stop_index: usize) -> Result<bool, GraphError>;
+
+    /// Returns the length of the shortest path (in number of nodes) from a start to a stop index.
+    fn shortest_path_len(
+        &self,
+        start_index: usize,
+        stop_index: usize,
+    ) -> Result<Option<usize>, GraphError>;
+
+    /// Finds the complete shortest path from a start to a stop index.
+    fn shortest_path(
+        &self,
+        start_index: usize,
+        stop_index: usize,
+    ) -> Result<Option<Vec<usize>>, GraphError>;
+
+    /// Finds the shortest path in a weighted graph using Dijkstra's algorithm.
+    ///
+    /// The edge weight type `W` must support addition, comparison, and have a zero value.
+    ///
+    /// # Returns
+    /// A tuple containing the sequence of node indices in the path and the total cost of that path.
+    /// Returns `None` if no path exists.
+    fn shortest_weighted_path(
+        &self,
+        start_index: usize,
+        stop_index: usize,
+    ) -> Result<Option<(Vec<usize>, W)>, GraphError>
+    where
+        W: Copy + Ord + Default + std::ops::Add<Output = W>;
+}
